+++ ui/command_test.go
@@ -13,7 +13,39 @@ var commandTests = []printTestCase{
 		func() { PrintCommand("echo hello") },
 		VerbosityLevelQuiet,
 		VerbosityLevelNormal,
-		fmt.Sprintf("[%s] %s\n", commandActionString, "echo hello"),
+		fmt.Sprintf("[%s] %s\n", runningString, "echo hello"),
+	},
+	{
+		`PrintEnvironment()`,
+		LoggerStderr,
+		func() {
+			a := "one"
+			c := "three"
+
+			PrintEnvironment(map[string]*string{
+				"A": &a,
+				"B": nil,
+				"C": &c,
+			})
+		},
+		VerbosityLevelQuiet,
+		VerbosityLevelNormal,
+		fmt.Sprintf(
+			"[%s] %s\n%sset %s=%s\n%sset %s=%s\n%sunset %s\n",
+			runningString,
+			environmentMessage,
+			outputPrefix, "A", "one",
+			outputPrefix, "C", "three",
+			outputPrefix, "B",
+		),
+	},
+	{
+		`PrintEnvironment(nil)`,
+		LoggerStderr,
+		func() { PrintEnvironment(nil) },
+		VerbosityLevelQuiet,
+		VerbosityLevelNormal,
+		"",
 	},
 	{
 		`PrintSkipped("echo hello", "oops")`,